+++ game/add.php
@@ -1,6 +1,7 @@
 <!doctype php>
 <?php
-	require_once "server/initialize.php";
+	$URLDepth = '../';
+	require_once "../server/initialize.php";
 	if(!empty($_POST)){
 	//	p
 		$_POST = array_map("strip_tags", $_POST);
